feat(observability): trim OTel footprint + native OTLP metric push (RFC-0019)
Phase 3 (footprint) and phase 4 (OTLP metrics) of RFC-0019.

- Drop autoprop for an explicit W3C TraceContext+Baggage composite propagator,
  removing the aws/b3/jaeger/ot propagator modules from the build. OTEL_PROPAGATORS
  no longer selects non-W3C propagators (documented trade-off); the chart value is
  retained for schema compatibility.
- Bump semconv v1.4.0 -> v1.41.0.
- Add an opt-in OTLP periodic metric reader (OTEL_METRICS_EXPORTER contains
  "otlp"), attached alongside the always-on Prometheus bridge so the scrape is
  never replaced. New openTelemetry.metricsExporter chart value (default
  "prometheus") wires OTEL_METRICS_EXPORTER into every component.

+++ pkg/storagesvc/storagesvc.go
@@ -232,7 +232,7 @@ func (ss *StorageService) deleteHandler(w http.ResponseWriter, r *http.Request) 
 	}
 
 	totalArchives.Add(r.Context(), -1)
-	totalMemoryUsage.Add(r.Context(), -int64(filesize))
+	totalMemoryUsage.Add(r.Context(), -filesize)
 	w.WriteHeader(http.StatusOK)
 }
